Install xorg-xdpyinfo on arch too.
Since a new release was made to PyVirtualDisplay that included [a
commit][commit] which tries to run `xdpyinfo` it is [now][] failing
tests because of its absence. For two reasons:

* warning messages fail tests unless ignored with caplog
* DeprecationWarnings (from using `log.warn) may also fail tests

This wouldn't normally be a problem because it is only run when it is
set up which is done before the tests start running. Unfortunately the
integration between and coverage.py appears to be re-importing parts of
the test setup so now all the tests around coverage are breaking.

It looks like there might be ways to tweak how pytest and coverage.py
work together but that code looks a little convoluted in this is
probably the easiest fix.

[commit]: https://github.com/ponty/PyVirtualDisplay/commit/3b11cf1e1381921c047ba18f4c5d929847f23b84
[now]: https://github.com/qutebrowser/qutebrowser/pull/4812